some unit tests for the player
